Refactor test files and update documentation
Moved test-related files from src/ to test/ directory for better organization. Updated CLAUDE.md and README.md with improved architecture, configuration, and track examples. Changed main entry point in index.html to test/main.tsx. Switched from pnpm to yarn by removing pnpm lock/workspace files and adding yarn.lock.
