+++ CLAUDE.md
@@ -31,6 +31,7 @@ npm run preview
 ## Architecture Overview
 
 ### State Management
+
 - **Zustand stores**: Multiple specialized stores instead of single global state
 - **Store factories**: `createBrowserStore()` and `createTrackStore()` for creating instances
 - **React Context**: `BrowserContext.tsx` provides store access to components
@@ -41,6 +42,7 @@ npm run preview
   - UI stores: `modalStore.ts`, `tooltipStore.ts`, `contextMenuStore.ts`
 
 ### Component Architecture
+
 ```
 Browser (main entry point)
 ├── DataFetcher (Apollo GraphQL orchestration)
@@ -55,9 +57,11 @@ Browser (main entry point)
 ```
 
 ### Track System
+
 The library supports multiple genomic data visualization types:
+
 - **BigWig**: Continuous signal data (Full/Dense modes)
-- **BigBed**: Discrete genomic regions (Dense/Squish modes) 
+- **BigBed**: Discrete genomic regions (Dense/Squish modes)
 - **BulkBed**: Multiple BigBed datasets in one track
 - **Transcript**: Gene annotations (Pack/Squish modes)
 - **Motif**: DNA sequence motifs
@@ -67,15 +71,17 @@ The library supports multiple genomic data visualization types:
 Each track type has its own component directory under `src/components/tracks/[tracktype]/`.
 
 ### Data Fetching
+
 - **Apollo Client**: GraphQL-based data fetching in `src/api/`
 - **DataFetcher.tsx**: Main orchestration component that manages all track data requests
 - **Query builders**: Modular query construction in `src/api/requestBuilder.ts`
 - **Caching**: Zustand dataStore manages cached genomic data with loading states
 
 ### Key Directories
+
 - `src/components/browser/`: Main browser wrapper components
 - `src/components/tracks/`: Track-specific rendering components
-- `src/components/modal/`: Configuration modal dialogs 
+- `src/components/modal/`: Configuration modal dialogs
 - `src/components/tooltip/`: Interactive tooltip system
 - `src/store/`: Zustand state management stores
 - `src/hooks/`: Custom React hooks for browser interactions
@@ -83,12 +89,15 @@ Each track type has its own component directory under `src/components/tracks/[tr
 - `src/api/`: GraphQL data fetching infrastructure
 
 ### Library Entry Points
+
 - **Public API**: `src/lib.ts` exports all public components, types, and store factories
 - **Development**: `src/main.tsx` provides demo application for testing
 - **Build output**: ES modules in `dist/` directory with TypeScript declarations
 
 ### Browser Context Pattern
+
 Components access stores through React Context instead of direct imports:
+
 ```tsx
 const browserStore = useBrowserStore();
 const trackStore = useTrackStore();
@@ -97,13 +106,18 @@ const trackStore = useTrackStore();
 This allows multiple browser instances with isolated state in the same application.
 
 ### SVG-Based Rendering
+
 All visualization uses SVG for crisp, scalable graphics. Track components render directly to SVG elements with coordinate transformations handled by custom hooks (`useXTransform.ts`, `useBrowserScale.ts`).
 
 ### Testing and Development
+
 - **Vitest**: Unit testing framework
-- **Playwright**: End-to-end testing  
+- **Playwright**: End-to-end testing
 - **Storybook**: Component development and documentation
 - **TypeScript**: Strict type checking with separate build configs
 
 ### Special Instructions
-- Never ask to run `npm run dev`. I will test the library manually. You may run `npm run lint` but try to avoid it if you can.
+
+- Never ask to run `npm run dev`. I will test the library manually. You may run `npm run lint --quiet` but try to avoid it if you can.
+
+- Don't hallucinate random URI/URLs for the tracks. You may instead use an already utilized URL or put a placeholder string such as "YOUR_URL_HERE"
