+++ index.html
@@ -9,6 +9,6 @@
   <body>
     <div id="root"></div>
     <script src="https://unpkg.com/react-scan/dist/auto.global.js"></script>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/test/main.tsx"></script>
   </body>
 </html>
