+++ test/main.tsx
@@ -1,4 +1,4 @@
-import { StrictMode, useEffect } from "react";
+import React, { StrictMode, useEffect } from "react";
 import { createRoot } from "react-dom/client";
 import { create } from "zustand";
 import {
@@ -11,10 +11,12 @@ import {
   Track,
   TrackType,
   Transcript,
-} from "./lib";
-import { createBrowserStore, InitialBrowserState, BrowserStoreInstance } from "./store/browserStore";
-import { TrackStoreInstance } from "./store/trackStore";
-import { Vibrant } from "./utils/color";
+  TrackStoreInstance,
+  createBrowserStore,
+  InitialBrowserState,
+  BrowserStoreInstance,
+  Vibrant,
+} from "../src/lib";
 import { bigBedExample, bigWigExample, bulkBedExample, motifExample, transcriptExample } from "./tracks";
 
 const useStore = create<{ name: string; setName: (name: string) => void }>((set) => ({
@@ -114,6 +116,23 @@ function Main() {
   ];
 
   const trackStore = createTrackStore(tracks);
+  const insertTrack = trackStore((state) => state.insertTrack);
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      insertTrack({
+        id: "delayed-bigwig",
+        title: "Delayed BigWig Track",
+        titleSize: 12,
+        height: 75,
+        color: Vibrant[1],
+        trackType: TrackType.BigWig,
+        url: "https://downloads.wenglab.org/DNAse_All_ENCODE_MAR20_2024_merged.bw",
+        displayMode: DisplayMode.Full,
+      });
+    }, 3000);
+
+    return () => clearTimeout(timer);
+  }, []);
 
   return (
     <div style={{ display: "flex", flexDirection: "column", alignItems: "center" }}>
